Update Gradle Wrapper to 8.11.1 and Android Gradle Plugin to 8.9.1
This commit updates the Gradle Wrapper version to 8.11.1 in `gradle-wrapper.properties` and upgrades the Android Gradle Plugin to version 8.9.1 in the top-level `build.gradle` file. Additionally, it disables minification for debug builds by adding `minifyEnabled false` and application id suffix to  `.dev` to debug builds in the app module's `build.gradle`.

Signed-off-by: Mahmoud Habib <mhmoudrhabib@gmail.com>